♻️  status_responses.py 추출 — 6 read-only responders (P0-L step 5)
변경 이유
- `format_status_diagnostic_response` 단독 450 줄 + 5 신규 read-only
  responder 가 한 파일에 섞여 있어 추가 surface (예: vault audit, growth
  ledger 라인) 가 elif 사슬 깊이를 키우고 있었다.
  하나의 모듈로 분리해 향후 status 라인 추가 시 변경 표면을 좁힌다.

주요 변경 사항
- 새 파일 `engineering_conversation/status_responses.py` (약 720 줄):
  - 6 public responder: `format_status_diagnostic_response`,
    `format_session_count_response`, `format_session_list_response`,
    `format_blocked_reason_response`, `format_continue_existing_response`,
    `format_change_direction_response`
  - private helper: `_open_states_set`, `_safe_list_sessions`,
    `_coerce_str`, `_format_coding_status_line`
  - 상수: `_STATUS_SEVERITY_TAGS`, `_MEMBER_BOT_PHRASES`
  - 헬퍼: `_asks_about_member_bots`
- `_legacy.py` 의 status 관련 본문 3 블록 (172 + 38 + 483 = 693 줄)
  제거 후 `from .status_responses import ...` 로 re-export 만 유지.
  `_format_general_help` 는 response_formatters 책임이라 보존.
- `__init__.py` facade 가 status responders 도 `.status_responses` 직접
  import 으로 끌어옴 — `_legacy` 의존도가 한 단계 더 줄었다.

비고
- 4657 PASS + 1 skipped 회귀 유지 (동작 변경 0).
- `_asks_about_member_bots` 가 `_normalize` 를 lazy `from ._legacy`
  import 으로 사용 — 다음 step 에서 intent_detection 추출 후 정리.
- 다음 step (6): intent_detection.py 추출 (가장 큰 책임 단위).
